requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.11.8
cffi==1.14.4
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
Django==3.1.3
django-rest-auth==0.9.5
django-rq==2.4.0
djangorestframework==3.12.2
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
msgpack==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.0
pytz==2020.4
redis==3.5.3
requests==2.25.0
rq==1.7.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.26.2
wincertstore==0.2
zope.interface==5.2.0
